refactor: crear package de películas/filmes
