+++ app/src/main/res/navigation/star_wars_navigation.xml
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/filmsFragment"
-        android:name="com.banregio.devuapp.starwars.presentation.FilmsFragment"
+        android:name="com.banregio.devuapp.starwars.presentation.films.FilmsFragment"
         android:label="FilmsFragment"
         tools:layout="@layout/fragment_films" />
